yaml配置多数据库

+++ utils/lv_conv/convert.go
@@ -18,8 +18,8 @@ func SubStr(str string, startIndex, endIndex int) string {
 }
 
 // 将带分隔符的字符串切成int64数组
-func ToInt64Array(str, split string) []int64 {
-	result := make([]int64, 0)
+func ToIntArray(str, split string) []int {
+	result := make([]int, 0)
 	if str == "" {
 		return result
 	}
@@ -27,7 +27,7 @@ func ToInt64Array(str, split string) []int64 {
 	if len(arr) > 0 {
 		for i := range arr {
 			if arr[i] != "" {
-				result = append(result, cast.ToInt64(arr[i]))
+				result = append(result, cast.ToInt(arr[i]))
 			}
 		}
 	}
